updating volume entity to return int64 for volume. (#115)

+++ alpaca/entities.go
@@ -134,7 +134,7 @@ type Bar struct {
 	High   float32 `json:"h"`
 	Low    float32 `json:"l"`
 	Close  float32 `json:"c"`
-	Volume int32   `json:"v"`
+	Volume int64   `json:"v"`
 }
 
 type ListBarParams struct {
